Bump @capacitor/core to v2.4.2
