--- android/capacitor-auth0/settings.gradle
@@ -1,2 +1,0 @@
-include ':capacitor-android'
-project(':capacitor-android').projectDir = new File('../node_modules/@capacitor/android/capacitor')
