+++ android/settings.gradle
@@ -1,0 +1,2 @@
+include ':capacitor-android'
+project(':capacitor-android').projectDir = new File('../node_modules/@capacitor/android/capacitor')
