+++ android/src/main/res/values/colors.xml
