+++ android/src/main/res/values/styles.xml
