+++ ios/.gitignore
@@ -1,1 +1,2 @@
-/Plugin/Pods/
+/Pods/
+xcuserdata/
