+++ ios/Plugin/Info.plist
