--- ios/Plugin/Podfile
@@ -1,18 +1,0 @@
-# Uncomment the next line to define a global platform for your project
-platform :ios, '11.0'
-
-target 'Plugin' do
-  # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
-  use_frameworks!
-
-  pod 'Auth0', '~> 1.14'
-
-  # Pods for IonicRunner
-  pod 'Capacitor'
-end
-
-target 'PluginTests' do
-  use_frameworks!
-
-  pod 'Capacitor'
-end
