--- rollup.config.js
@@ -1,14 +1,0 @@
-import nodeResolve from 'rollup-plugin-node-resolve';
-
-export default {
-  input: 'dist/esm/index.js',
-  output: {
-    file: 'dist/plugin.js',
-    format: 'iife',
-    name: 'capacitorPlugin',
-    sourcemap: true
-  },
-  plugins: [
-    nodeResolve()
-  ]
-};
